+++ src/addons/character_dna/release_notes.md
@@ -1,7 +1,13 @@
+## Minor Changes
+
+* Added transfer to shape key operator in Raw Editor
+
 ## Patch Changes
 
 * Made DNA component type detection more robust [#337](https://github.com/poly-hammer/character-dna-addon/issues/337)
 * Fixed body animation pose mode switching bug
+* Fixed view options panel lagging draw calls
+* Fixed operator for adding [Texture Logic Node](https://docs.polyhammer.com/character-dna-addon/terminology/#texture-logic) now under `Add > Rig Logic > Head Texture Logic` in material graph
 
 ## Tests Passing On